chore: split ProjectCard into Legacy and New behind newProjectList flag
